add: config.main.ts for app and e2e tests
