+++ test/app.e2e-spec.ts
@@ -2,6 +2,7 @@ import { Test, TestingModule } from '@nestjs/testing';
 import { INestApplication } from '@nestjs/common';
 import request from 'supertest';
 import { AppModule } from '../src/app.module';
+import { configNestApp } from '../src/config.main';
 
 describe('AppController (e2e)', () => {
     let app: INestApplication;
@@ -11,7 +12,8 @@ describe('AppController (e2e)', () => {
             imports: [AppModule],
         }).compile();
 
-        app = moduleFixture.createNestApplication();
+        const baseApp = moduleFixture.createNestApplication();
+        app = configNestApp(baseApp);
         await app.init();
     });
